+++ assets/javascripts/kitten/components/buttons/button/helpers/modifier-styles.js
@@ -67,7 +67,7 @@ export const modifierStyles = modifier => {
       break
 
     case 'boron':
-      borderColor = COLORS.line1
+      borderColor = COLORS.grey1
       backgroundColor = COLORS.font1
       color = COLORS.background1
       hoverBorderColor = 'currentColor'