+++ src/constants/colors-config.js
@@ -25,6 +25,8 @@ var _default = {
   // Grey
   line3: '#2d2d2d',
   // Extra semi dark grey, not used for text
+  grey1: '#636363',
+  // Medium grey used for text or not
   primary1: '#19b4fa',
   // Primary
   primary2: '#05a8e6',