+++ src/esm/components/layout/dashboard-layout/index.js
@@ -13,6 +13,7 @@ import { getReactElementsByType, getReactElementsWithoutTypeArray } from '../../
 import { DASHBOARD_HIDE_CONTENT_EVENT, DASHBOARD_SHOW_CONTENT_EVENT, dispatchEvent } from '../../../helpers/dom/events';
 import { BurgerIcon } from '../../../components/icons/burger-icon';
 import { ArrowIcon } from '../../../components/icons/arrow-icon';
+import { Alert as AlertComponent } from '../../../components/notifications/alert';
 import { Flow } from './flow';
 import { StyledDashboard } from './styles';
 export var DashboardLayout = function DashboardLayout(_ref) {
@@ -194,14 +195,17 @@ export var DashboardLayout = function DashboardLayout(_ref) {
       onClick: handleButtonClick,
       'aria-expanded': isOpen ? isOpen : null
     })
-  }), /*#__PURE__*/React.createElement("main", {
+  }), renderComponentChildrenArray(getReactElementsByType({
+    children: children,
+    type: Alert
+  })), /*#__PURE__*/React.createElement("main", {
     className: classNames('k-DashboardLayout__main', {
       'k-DashboardLayout__main--fullHeight': fullHeightContent
     }),
     id: "main"
   }, renderComponentArray(getReactElementsWithoutTypeArray({
     children: children,
-    typeArray: [SiteHeader, Header, SideContent, SideFooter]
+    typeArray: [SiteHeader, Header, SideContent, SideFooter, Alert]
   }))))));
 };
 
@@ -263,6 +267,15 @@ var SideFooter = function SideFooter(_ref6) {
   }, props));
 };
 
+var Alert = function Alert(_ref7) {
+  var className = _ref7.className,
+      props = _objectWithoutProperties(_ref7, ["className"]);
+
+  return /*#__PURE__*/React.createElement(AlertComponent, _extends({
+    className: classNames('k-DashboardLayout__alert', 'k-DashboardLayout__fullWidth', className)
+  }, props));
+};
+
 DashboardLayout.propTypes = {
   backLinkProps: PropTypes.object,
   buttonProps: PropTypes.shape({
@@ -285,3 +298,4 @@ DashboardLayout.Header = Header;
 DashboardLayout.SideContent = SideContent;
 DashboardLayout.SideFooter = SideFooter;
 DashboardLayout.Flow = Flow;
+DashboardLayout.Alert = Alert;