+++ src/index.js
@@ -1367,6 +1367,12 @@ Object.defineProperty(exports, "LockIcon", {
     return _lockIcon.LockIcon;
   }
 });
+Object.defineProperty(exports, "LockOutlineIcon", {
+  enumerable: true,
+  get: function get() {
+    return _lockOutlineIcon.LockOutlineIcon;
+  }
+});
 Object.defineProperty(exports, "LoginIcon", {
   enumerable: true,
   get: function get() {
@@ -1431,6 +1437,12 @@ Object.defineProperty(exports, "QuestionMarkIcon", {
   enumerable: true,
   get: function get() {
     return _questionMarkIcon.QuestionMarkIcon;
+  }
+});
+Object.defineProperty(exports, "RefundIcon", {
+  enumerable: true,
+  get: function get() {
+    return _refundIcon.RefundIcon;
   }
 });
 Object.defineProperty(exports, "RocketCircleIcon", {
@@ -1587,6 +1599,12 @@ Object.defineProperty(exports, "ShieldIcon", {
   enumerable: true,
   get: function get() {
     return _shieldIcon.ShieldIcon;
+  }
+});
+Object.defineProperty(exports, "ShieldCheckIcon", {
+  enumerable: true,
+  get: function get() {
+    return _shieldCheckIcon.ShieldCheckIcon;
   }
 });
 Object.defineProperty(exports, "PeopleIcon", {
@@ -2318,6 +2336,8 @@ var _locationIcon = require("./components/icons/location-icon");
 
 var _lockIcon = require("./components/icons/lock-icon");
 
+var _lockOutlineIcon = require("./components/icons/lock-outline-icon");
+
 var _loginIcon = require("./components/icons/login-icon");
 
 var _maestroIcon = require("./components/icons/maestro-icon");
@@ -2339,6 +2359,8 @@ var _phoneIcon = require("./components/icons/phone-icon");
 var _pinterestIcon = require("./components/icons/pinterest-icon");
 
 var _questionMarkIcon = require("./components/icons/question-mark-icon");
+
+var _refundIcon = require("./components/icons/refund-icon");
 
 var _rocketCircleIcon = require("./components/icons/rocket-circle-icon");
 
@@ -2391,6 +2413,8 @@ var _fileIcon = require("./components/icons/file-icon");
 var _loudspeakerIcon = require("./components/icons/loudspeaker-icon");
 
 var _shieldIcon = require("./components/icons/shield-icon");
+
+var _shieldCheckIcon = require("./components/icons/shield-check-icon");
 
 var _peopleIcon = require("./components/icons/people-icon");